==> TitanicMachineLearningfromDisaster/1_SimpleAnalysis.ipynb <==
### 导入数据

import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')
%matplotlib inline

train = pd.read_csv('data/train.csv') # 读取数据
test = pd.read_csv('data/test.csv')
test1 = pd.read_csv('data/test.csv')

### 观察数据

train.head()

train.info()

train.describe()

### 数据预处理

1、观察数据后发现，可去除PassengerId，Name，Ticket（无关）；Cabin（缺失过多）几列特征

train = train.drop(['PassengerId', 'Name', 'Ticket', 'Cabin'], axis=1)
train.head()

2、至此Age字段和Embarked字段有缺失，需补充

train['Age'] = train['Age'].fillna(train['Age'].median()) # 由于Age最大值最小值差距过大，这里用中位数进行填充

train['Embarked'].value_counts() # 统计Embarked各类别出现次数

train['Embarked'] = train['Embarked'].fillna('S') # 用出现最多的类别填充Embarked列
train.info()

3、将Sex和Embarked由字符串转化为数字

from sklearn.preprocessing import LabelEncoder

train['Sex'] = LabelEncoder().fit_transform(train['Sex'])
train['Embarked'] = LabelEncoder().fit_transform(train['Embarked'])

train.info()

4、将数据做标准化处理

from sklearn import preprocessing

scaler = preprocessing.StandardScaler()
for feature in train.columns:
    if feature != 'Survived':
        train[feature] = scaler.fit_transform(train[feature].values.reshape(-1, 1)) # 标准化处理数据

train.head()

### 使用逻辑回归预测

from sklearn.model_selection import cross_val_score
from sklearn.linear_model import LogisticRegression

x_train = train.drop(['Survived'], axis=1)
y_train = train['Survived']

C_params = [0.01, 0.1, 1, 10, 100] # 参数C,惩罚力度

for C_param in C_params:
    print('========================================')
    print('C_param = ', C_param)
    lr =  LogisticRegression(C=C_param, penalty='l1') # 正则化项采用l1范数
    score = cross_val_score(lr, x_train, y_train, cv=5) # 采用k折交叉验证，cv为折树
    print('score：',score.mean()) # 当前参数下交叉验证模型评分

可见使用C=0.1时分数最高，score为0.786

### 使用随机森林预测

from sklearn.model_selection import cross_val_score
from sklearn.ensemble import RandomForestClassifier

dt =RandomForestClassifier(n_estimators = 100,min_samples_split=4,min_samples_leaf=2)
score = cross_val_score(dt,x_train, y_train, cv=10)
print('score：',score.mean())

可见使用随机森林预测后准确率大大增加

from sklearn import preprocessing
from sklearn.preprocessing import LabelEncoder

test = test.drop(['PassengerId', 'Name', 'Ticket', 'Cabin'], axis=1)
test['Age'] = test['Age'].fillna(test['Age'].median())
test['Fare'] = test['Fare'].fillna(test['Fare'].median())
test['Embarked'] = test['Embarked'].fillna('S')

test['Sex'] = LabelEncoder().fit_transform(test['Sex'])
test['Embarked'] = LabelEncoder().fit_transform(test['Embarked'])

scaler = preprocessing.StandardScaler()
for feature in train.columns:
    if feature != 'Survived':
        test[feature] = scaler.fit_transform(test[feature].values.reshape(-1, 1)) # 标准化处理数据

test.head()

test['Fare'] = test['Fare'].fillna(test['Fare'].median())

dt.fit(x_train, y_train)
predict_data = dt.predict(test)

submission = pd.DataFrame({
    'PassengerId':test1['PassengerId'],
    'Survived':predict_data
})
submission.to_csv('gender_submission.csv', index=False)

提交之后分数为0.75598，有些糟糕
