==> TitanicMachineLearningfromDisaster/3_DataPreprocessing.ipynb <==
import numpy as np
import pandas as pd
import warnings

warnings.filterwarnings('ignore')

train_data = pd.read_csv('data/train.csv')
test_data = pd.read_csv('data/test.csv')

print(train_data.info()) # 查看数去条数方便数据完成后分离      
print('*'*30)
print(test_data.info())

# 将训练数据和测试数据组合在一起方便处理
train = train_data.loc[:,train_data.columns != 'Survived'] # 去掉标签
test = test_data
data = pd.concat([train,test], ignore_index=True) # 合并并整理index
data.head()

# 去调PassengerId, Ticket
data.drop(['PassengerId', 'Ticket'], axis=1, inplace=True)
data.head()

# 处理Pclass列
pclass_dummies = pd.get_dummies(data['Pclass'])
pclass_dummies.columns = ['Pclass1', 'Pclass2', 'Pclass3']
data = data.join(pclass_dummies)
data.drop(['Pclass'], axis=1, inplace=True)
data.head()

# 处理Embarked列
embarked_dummies = pd.get_dummies(data['Embarked'])
embarked_dummies.columns = ['Embarked1', 'Embarked2', 'Embarked3']
data = data.join(embarked_dummies)
data.drop(['Embarked'], axis=1, inplace=True)
data.head()

# 提取Name中的称呼
import re

title = data['Name'].map(lambda x: re.compile(", (.*?)\.").findall(x)[0])

title_Dict = {}
title_Dict.update(dict.fromkeys(['Capt', 'Col', 'Major', 'Dr', 'Rev'], 'Officer'))
title_Dict.update(dict.fromkeys(['Don', 'Sir', 'the Countess', 'Dona', 'Lady'], 'Royalty'))
title_Dict.update(dict.fromkeys(['Mme', 'Ms', 'Mrs'], 'Mrs'))
title_Dict.update(dict.fromkeys(['Mlle', 'Miss'], 'Miss'))
title_Dict.update(dict.fromkeys(['Mr'], 'Mr'))
title_Dict.update(dict.fromkeys(['Master','Jonkheer'], 'Master'))

data['Title'] = title.map(title_Dict)
data.drop(['Name'], axis=1,inplace=True)
data.head()

# 处理Title列
title = pd.get_dummies(data['Title'])
data = data.join(title)

data.drop(['Title'], axis=1, inplace=True)
data.head()

# 处理Sex列
from sklearn.preprocessing import LabelEncoder

data['Sex'] = LabelEncoder().fit_transform(data['Sex'])
data.head()

# 处理Age列
from sklearn import preprocessing

scaler = preprocessing.StandardScaler()
data['Age'].fillna(data['Age'].median(), inplace=True)
data['Age'] = scaler.fit_transform(data['Age'].values.reshape(-1, 1)) # 标准化处理数据
data.head()

# 处理SibSp和Parch列

def family_size_category(family_size):
    if family_size <= 1:
        return 'Single'
    elif family_size <= 4:
        return 'Small_Family'
    else:
        return 'Large_Family'
    
data['Family_Size'] = data['Parch'] + data['SibSp'] + 1
data['Family_Size'] = data['Family_Size'].map(family_size_category)
data.head()

family_size = pd.get_dummies(data['Family_Size'])
data = data.join(family_size)
data.drop(['Family_Size'], axis=1, inplace=True)
data.head()

data.drop(['SibSp', 'Parch'], axis=1, inplace=True)

# 处理票价

from sklearn import preprocessing

scaler = preprocessing.StandardScaler()
data['Fare'].fillna(data['Fare'].mean(), inplace=True)
data['Fare'] = scaler.fit_transform(data['Fare'].values.reshape(-1, 1)) # 标准化处理数据
data.head()

# Cabin缺失过多，去掉
data.drop(['Cabin'], axis=1, inplace=True)

data.info()

# Age用中位数填充
data['Age'] = data['Age'].fillna(data['Age'].median())

# Fare用平均数填充
data['Fare'] = data['Fare'].fillna(data['Fare'].mean())

# 将训练数据集和测试数据集分开

train_data = data.loc[:890]
train_data['Survived'] = pd.read_csv('data/train.csv')['Survived']
test_data = data.loc[891:].reset_index(drop=True)
test_data['PassengerId'] = pd.read_csv('data/test.csv')['PassengerId']

print(train_data.info())   
print('*'*30)
print(test_data.info())

# 保存train_data和test_data
train_data.to_csv('data/train_data.csv', index=False)
test_data.to_csv('data/test_data.csv', index=False)
