==> TitanicMachineLearningfromDisaster/4_Predict.ipynb <==
import numpy as np
import pandas as pd
import warnings

warnings.filterwarnings('ignore')

train_data = pd.read_csv('data/train_data.csv')
test_data = pd.read_csv('data/test_data.csv')

x_train = train_data.loc[:,(train_data.columns != 'Survived')]
y_train = train_data['Survived']

<font color=blue>训练几个分类器查看效果</font>

# 逻辑回归
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import GridSearchCV

penalty = ['l1', 'l2']
C = [0.001, 0.01, 0.1, 1, 10, 100, 1000]
parameters = dict(penalty=penalty, C=C) # 参数

lr = LogisticRegression()
grid = GridSearchCV(lr, parameters, cv=5)
grid.fit(x_train, y_train)
# grid.cv_results_
print('best params:', grid.best_params_)
print('best score:', grid.best_score_)

# 支持向量机
from sklearn.svm import SVC
from sklearn.model_selection import GridSearchCV

gamma = [0.001, 0.01, 0.1, 1, 10, 100, 100]
C = [0.001, 0.01, 0.1, 1, 10, 100, 100]
parameters = dict(gamma=gamma, C=C) # 参数

svm = SVC()
grid = GridSearchCV(svm, parameters, cv=5)
grid.fit(x_train, y_train)

print('best params:', grid.best_params_)
print('best score:', grid.best_score_)

# 随机森林
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import GridSearchCV

n_estimators = [10, 15, 20] # 弱分类器个数
criterion = ['gini', 'entropy'] # 划分时评价标准
min_samples_leaf = [2, 4, 6] # 叶子节点最小样本数
parameters = dict(n_estimators=n_estimators, criterion=criterion, min_samples_leaf=min_samples_leaf)

clf = RandomForestClassifier()
grid = GridSearchCV(clf, parameters, cv=5)
grid.fit(x_train, y_train)

print('best params:', grid.best_params_)
print('best score:', grid.best_score_)

# 使用随机森林预测

clf = RandomForestClassifier(n_estimators=10, criterion='gini', min_samples_leaf=2)
clf.fit(x_train, y_train)
predict_data = clf.predict(test_data.loc[:, test_data.columns != 'PassengerId'])

submission = pd.DataFrame({
    'PassengerId':test_data['PassengerId'],
    'Survived':predict_data
})

submission.to_csv('data/gender_submission.csv', index=False)

<font color=blue>得分0.75598，没有简单的决策树分数高</font>
